// frontend/src/client/types.gen.ts
export type FeatureResponse = {
  /**
   * Name
   */
  name: string;
  /**
   * Code
   */
  code: string;
  /**
   * Unit Price
   */
  unit_price: string;
  /**
   * Max Unit Limit
   */
  max_unit_limit: number;
  /**
   * Is Active
   */
  is_active?: boolean;
  /**
   * Id
   */
  id: number;
};
// ...
export type UserResponse = {
  /**
   * Name
   */
  name: string;
  /**
   * Email
   */
  email: string;
  role?: UserRole;
  /**
   * Id
   */
  id: number;
  /**
   * Profile Img
   */
  profile_img?: string | null;
};

/**
 * UserRole
 */
export type UserRole = "admin" | "buyer";